Fix a number of warnings revealed by Linux compilers.
Tried it with:

- g++ (Ubuntu 5.4.0-6ubuntu1~16.04.5) 5.4.0 20160609
- g++-6 (Ubuntu/Linaro 6.3.0-18ubuntu2~16.04) 6.3.0 20170519
- clang version 6.0.0-svn321396-1~exp1 (trunk)

+++ src/b2/load_save.cpp
@@ -366,7 +366,7 @@ static bool LoadFile2(ContType *data,
     if(len<0) {
         AddError(msg,path,"load","ftell failed",errno);
         goto done;
-    } else if((size_t)len>SIZE_MAX) {
+    } else if(len>SIZE_MAX) {
         AddError(msg,path,"load","file is too large",0);
         goto done;
     }
